+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## Unreleased
+
+
+## [0.0.4] - 2024-11-04
 ### Added
 - Return slices as new AffineCipher instances
 - New methods implemented for AffineCipher
@@ -36,7 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial public release
 
 
-[Unreleased]: https://github.com/jfolz/shufflish/compare/0.0.3...main
+[Unreleased]: https://github.com/jfolz/shufflish/compare/0.0.4...main
+[0.0.4]: https://github.com/jfolz/shufflish/compare/0.0.3...0.0.4
 [0.0.3]: https://github.com/jfolz/shufflish/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/jfolz/shufflish/compare/0.0.1...0.0.2
 [0.0.1]: https://github.com/jfolz/shufflish/releases/tag/0.0.1